LPF para angulo e centroid

+++ utils.cpp
@@ -227,11 +227,12 @@ pair <double, double> Low_pass_filter_coeff (float cutoff)
 Point Low_pass_filter_Centroid(Point centroid, Point last_centroid, Point last_proc_centroid, pair<double, double> coeff)
 {
 
-    int x;
-    int y;
+    double x;
+    double y;
 
-    x = (int) coeff.second * centroid.x + coeff.second * last_centroid.x + coeff.first * last_proc_centroid.x;
-    y = (int) coeff.second * centroid.y + coeff.second * last_centroid.y + coeff.first * last_proc_centroid.y;
+    x = coeff.second * centroid.x + coeff.second * last_centroid.x + coeff.first * last_proc_centroid.x;
+    y = coeff.second * centroid.y + coeff.second * last_centroid.y + coeff.first * last_proc_centroid.y;
+
     Point centroid_proc = Point(x, y);
 
     return centroid_proc;
